new solve file. and renamed the existing file for managing the grids.
