+++ src/ManagingGrid/solve.ts
@@ -1,0 +1,4 @@
+
+const solveGrid = (gridToSolve: string[][]): string[][] => {
+    return null
+}
